moves livelabel layout to base compound component
