--- src/app/components/LiveLabel/LiveLabelContext.tsx
@@ -1,4 +1,0 @@
-import React from 'react';
-
-const LiveLabelContext = React.createContext({});
-export default LiveLabelContext;
