+++ src/app/components/LiveLabel/index.test.tsx
@@ -1,39 +1,36 @@
 import React from 'react';
 import { render } from '../react-testing-library-with-providers';
-import LiveLabelPromo from '.';
+import LiveLabel from '.';
 
 describe('LiveLabel', () => {
   it('should render correctly with localised live text', () => {
-    const { container } = render(<LiveLabelPromo />, {
+    const { container } = render(<LiveLabel />, {
       service: 'pidgin',
     });
     expect(container).toMatchSnapshot();
   });
 
   it('should render correctly with English live text', () => {
-    const { container } = render(<LiveLabelPromo />, {
+    const { container } = render(<LiveLabel />, {
       service: 'russian',
     });
     expect(container).toMatchSnapshot();
   });
 
   it('should render correctly with English live text and children', () => {
-    const { container } = render(
-      <LiveLabelPromo> this is a headline </LiveLabelPromo>,
-      {
-        service: 'news',
-      },
-    );
+    const { container } = render(<LiveLabel> this is a headline </LiveLabel>, {
+      service: 'news',
+    });
     expect(container).toMatchSnapshot();
   });
 
   it('should render correctly with custom offscreen text', () => {
-    const { container } = render(<LiveLabelPromo offScreenText="Watch Live" />);
+    const { container } = render(<LiveLabel offScreenText="Watch Live" />);
     expect(container).toMatchSnapshot();
   });
 
   it('should correctly render for RTL service', () => {
-    const { container } = render(<LiveLabelPromo />, {
+    const { container } = render(<LiveLabel />, {
       service: 'arabic',
     });
     expect(container).toMatchSnapshot();
@@ -60,9 +57,7 @@ describe('LiveLabel', () => {
         expectedScreenReaderText,
       }) => {
         const { getByRole } = render(
-          <LiveLabelPromo offScreenText={offScreenText}>
-            {children}
-          </LiveLabelPromo>,
+          <LiveLabel offScreenText={offScreenText}>{children}</LiveLabel>,
           {
             service,
           },
