+++ src/app/legacy/psammead/psammead-story-promo/src/index.test.jsx
@@ -5,7 +5,7 @@ import styled from '@emotion/styled';
 import { GEL_SPACING_HLF } from '#psammead/gel-foundations/src/spacings';
 import { render } from '../../../../components/react-testing-library-with-providers';
 import latin from '../../../../components/ThemeProvider/fontScripts/latin';
-import LiveLabelPromo from '../../../../components/LiveLabelPromo';
+import LiveLabel from '../../../../components/LiveLabel';
 import StoryPromo, { Headline, Summary, Link } from './index';
 import relatedItems from '../testHelpers/relatedItems';
 import IndexAlsosContainer from '../testHelpers/IndexAlsosContainer';
@@ -22,9 +22,9 @@ const Info = ({ promoType, isLive, alsoItems, promoHasImage }) => (
     >
       <Link href="https://www.bbc.co.uk/news">
         {isLive ? (
-          <LiveLabelPromo ariaHidden offScreenText="Live">
+          <LiveLabel ariaHidden offScreenText="Live">
             The live promo headline
-          </LiveLabelPromo>
+          </LiveLabel>
         ) : (
           'The headline of the promo'
         )}
